requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
jinja2>=3.1.0
python-dateutil>=2.8.2
pydantic>=2.0.0
python-dotenv>=1.0.0
